Refactor reference parsing
